feat: add auth

+++ django_todo/core/views.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import redirect
 from django.shortcuts import render
 
@@ -5,24 +6,29 @@ from .forms import TaskForm
 from .models import Task
 
 
-# Create your views here.
+@login_required
 def home(request):
-    tasks = Task.objects.all()
+    tasks = Task.objects.filter(user=request.user)
     return render(request, "pages/home.html", {"tasks": tasks, "form": TaskForm()})
 
 
+@login_required
 def add_task(request):
     if request.method == "POST":
         form = TaskForm(request.POST)
         if form.is_valid():
-            form.save()
+            task = form.save(commit=False)
+            task.user = request.user
+            task.save()
             return redirect("core:home")
     return redirect("core:home")
 
 
+@login_required
 def delete_task(request, task_id):
     if request.method == "POST":
         task = Task.objects.get(id=task_id)
-        task.delete()
+        if task.user == request.user:  # Security check
+            task.delete()
         return redirect("core:home")
     return redirect("core:home")
